Modified routes and changed text
