+++ src/app/assign-users/columns.tsx
@@ -1,0 +1,54 @@
+"use client";
+
+import { ColumnDef } from "@tanstack/react-table";
+import { Button } from "@/components/ui/button";
+import { useRouter } from "next/navigation";
+
+export type User = {
+  id: string;
+  name: string;
+  email: string;
+  date: string;
+};
+
+const HandleAssignUser = (row: User) => {
+  sessionStorage.setItem(
+    "assignedUser",
+    JSON.stringify({
+      name: row.name,
+      email: row.email,
+      id: row.id,
+    }),
+  );
+};
+
+export const columns: ColumnDef<User>[] = [
+  {
+    accessorKey: "name",
+    header: "Client Name",
+  },
+  {
+    accessorKey: "email",
+    header: "Email",
+  },
+  {
+    accessorKey: "date",
+    header: "Date Added",
+  },
+  {
+    id: "actions",
+    cell: ({ row }) => {
+      return (
+        <Button
+          className="text-white w-full px-2 py-1 text-lg bg-basic-blue hover:bg-hover-blue rounded flex justify-center items-center"
+          onClick={() => {
+            HandleAssignUser(row.original);
+            window.location.href = "/create-event";
+          }}
+        >
+          Assign
+        </Button>
+      );
+    },
+  },
+];
