+++ src/app/assign-users/data-table.tsx
@@ -39,7 +39,7 @@ export function DataTable<TData, TValue>({ columns, data }: DataTableProps<TData
                   <TableCell
                     className={
                       cell.column.id === "actions"
-                        ? "w-[100px] text-lg"
+                        ? "w-[200px] text-lg"
                         : cell.column.id === "date"
                           ? "w-[300px] text-lg"
                           : "w-[400px] text-lg"
